Update the script to check an overall status of the project/analysis rather than just look at the clustering
